refactor(type): organize type header responsibilities
