+++ Source/Portfolio/AI/BehaviorTree/Service/CBTService_UpdateAIIntentState.cpp
@@ -15,7 +15,6 @@
 
 #include "Type/CStateTypes.h"
 #include "Type/CHealthTypes.h"
-#include "Type/CHealthTypes.h"
 #include "Type/CEngageAssignmentTypes.h"
 #include "AI/Blackboard/CAIKey.h"
 #include "AI/Blackboard/CAIBlackboardValueHelper.h"
