+++ Source/Portfolio/Component/CCombatSignalSourceComponent.h
@@ -4,7 +4,6 @@
 #include "Components/ActorComponent.h"
 #include "Type/CCharacterComponentReferenceTypes.h"
 #include "Type/CCombatSignalTypes.h"
-#include "Type/CCombatSignalTypes.h"
 #include "Type/CCombatHitTypes.h"
 #include "Type/CCombatDamageTypes.h"
 #include "Type/CCombatSignalSourceTypes.h"
