+++ Source/Portfolio/Component/CCombatSignalTargetComponent.h
@@ -4,7 +4,6 @@
 #include "Components/ActorComponent.h"
 #include "Type/CCharacterComponentReferenceTypes.h"
 #include "Type/CCombatSignalTypes.h"
-#include "Type/CCombatSignalTypes.h"
 #include "Type/CCombatDamageTypes.h"
 #include "Type/CCombatResultTypes.h"
 #include "Type/CCombatSignalTargetTypes.h"
