+++ Source/Portfolio/Controller/CAIController.h
@@ -3,6 +3,7 @@
 #include "CoreMinimal.h"
 #include "AIController.h"
 #include "Perception/AIPerceptionTypes.h"
+#include "Core/Debug/FAIPerceptionDebugTypes.h"
 #include "Type/CAITypes.h"
 #include "CAIController.generated.h"
