+++ Source/Portfolio/System/Combat/CWorldSubsystem_CombatEngage.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Subsystems/WorldSubsystem.h"
+#include "Core/Debug/FCombatEngageDebugTypes.h"
 #include "Type/CEngageAssignmentTypes.h"
 #include "CWorldSubsystem_CombatEngage.generated.h"
