+++ Source/Portfolio/Type/CAITypes.h
@@ -6,6 +6,8 @@
 #include "Type/CEngageAssignmentTypes.h"
 #include "CAITypes.generated.h"
 
+// Enum
+
 UENUM(BlueprintType)
 enum class EPatrolMode : uint8
 {
@@ -31,96 +33,32 @@ enum class EContextBuildResult : uint8
     Error
 };
 
-struct FPerceptionCandidateAuditState
+// Runtime State
+
+USTRUCT(BlueprintType)
+struct FPatrolPointData
 {
-    bool bEnabled = false;
+    GENERATED_BODY()
 
-    float RuntimeStartTime = 0.f;
-    uint64 RuntimeStartFrame = 0;
+    UPROPERTY(Transient)
+    FVector Location = FVector::ZeroVector;
 
-    float FirstRawPerceptionTime = -1.f;
-    uint64 FirstRawPerceptionFrame = 0;
+    UPROPERTY(Transient)
+    float ExtraWaitTime = 0.f;
 
-    float FirstValidTargetTime = -1.f;
-    uint64 FirstValidTargetFrame = 0;
+    UPROPERTY(Transient)
+    bool bFaceOnArrive = false;
 
-    int32 RawPerceptionEventCount = 0;
-    int32 MaxTargetDataMapSize = 0;
+    UPROPERTY(Transient)
+    float FaceYaw = 0.f;
 
-    TSet<TWeakObjectPtr<class AActor>> RawPerceptionActors;
-    TSet<TWeakObjectPtr<class AActor>> ValidTargetProviderActors;
-    TSet<TWeakObjectPtr<class AActor>> InvalidTargetProviderActors;
+    UPROPERTY(Transient)
+    FName PointTag = NAME_None;
 
-    void Reset()
-    {
-        bEnabled = false;
-
-        RuntimeStartTime = 0.f;
-        RuntimeStartFrame = 0;
-
-        FirstRawPerceptionTime = -1.f;
-        FirstRawPerceptionFrame = 0;
-
-        FirstValidTargetTime = -1.f;
-        FirstValidTargetFrame = 0;
-
-        RawPerceptionEventCount = 0;
-        MaxTargetDataMapSize = 0;
-
-        RawPerceptionActors.Reset();
-        ValidTargetProviderActors.Reset();
-        InvalidTargetProviderActors.Reset();
-    }
-};
-
-struct FBlackboardEngageLatencyAuditState
-{
-    bool bEnabled = false;
-
-    float RuntimeStartTime = 0.f;
-    uint64 RuntimeStartFrame = 0;
-
-    float FirstPerceptionContextTime = -1.f;
-    uint64 FirstPerceptionContextFrame = 0;
-
-    float FirstBlackboardTargetTime = -1.f;
-    uint64 FirstBlackboardTargetFrame = 0;
-
-    float FirstEngageRequestTime = -1.f;
-    uint64 FirstEngageRequestFrame = 0;
-
-    float FirstEngageAssignmentTime = -1.f;
-    uint64 FirstEngageAssignmentFrame = 0;
-
-    TWeakObjectPtr<class AActor> FirstPerceptionTargetActor;
-    TWeakObjectPtr<class AActor> FirstBlackboardTargetActor;
-    TWeakObjectPtr<class AActor> FirstEngageRequestTargetActor;
-    TWeakObjectPtr<class AActor> FirstEngageAssignmentTargetActor;
-
-    void Reset()
-    {
-        bEnabled = false;
-
-        RuntimeStartTime = 0.f;
-        RuntimeStartFrame = 0;
-
-        FirstPerceptionContextTime = -1.f;
-        FirstPerceptionContextFrame = 0;
-
-        FirstBlackboardTargetTime = -1.f;
-        FirstBlackboardTargetFrame = 0;
-
-        FirstEngageRequestTime = -1.f;
-        FirstEngageRequestFrame = 0;
-
-        FirstEngageAssignmentTime = -1.f;
-        FirstEngageAssignmentFrame = 0;
-
-        FirstPerceptionTargetActor.Reset();
-        FirstBlackboardTargetActor.Reset();
-        FirstEngageRequestTargetActor.Reset();
-        FirstEngageAssignmentTargetActor.Reset();
-    }
+public:
+    FPatrolPointData() = default;
+    FPatrolPointData(const FPatrolPointData&) = default;
+    FPatrolPointData& operator=(const FPatrolPointData&) = default;
 };
 
 USTRUCT(BlueprintType)
@@ -155,6 +93,8 @@ public:
         return IsValid(TargetActor);
     }
 };
+
+// Runtime Context
 
 USTRUCT(BlueprintType)
 struct FAIContext
@@ -218,32 +158,6 @@ public:
     {
         return IsValid(TargetActor);
     }
-};
-
-USTRUCT(BlueprintType)
-struct FPatrolPointData
-{
-    GENERATED_BODY()
-
-    UPROPERTY(Transient)
-    FVector Location = FVector::ZeroVector;
-
-    UPROPERTY(Transient)
-    float ExtraWaitTime = 0.f;
-
-    UPROPERTY(Transient)
-    bool bFaceOnArrive = false;
-
-    UPROPERTY(Transient)
-    float FaceYaw = 0.f;
-
-    UPROPERTY(Transient)
-    FName PointTag = NAME_None;
-
-public:
-    FPatrolPointData() = default;
-    FPatrolPointData(const FPatrolPointData&) = default;
-    FPatrolPointData& operator=(const FPatrolPointData&) = default;
 };
 
 USTRUCT(BlueprintType)
